+++ packages/genesys/packages/react-components/src/Label/Label.types.ts
@@ -2,8 +2,7 @@ import { CoreLabelProps } from "@peersyst/react-components-core";
 import { CSSProperties } from "react";
 import { Breakpoint } from "../theme";
 
-export interface LabelBreakpoint
-    extends Pick<CoreLabelProps, "placement" | "alignment" | "gap" | "singleLine"> {
+export interface LabelBreakpoint extends Pick<CoreLabelProps, "placement" | "alignment" | "gap"> {
     width: Breakpoint | number;
 }
 
@@ -20,4 +19,8 @@ export interface LabelProps extends CoreLabelProps {
      * Label style
      */
     style?: CSSProperties;
+    /**
+     * Text doesn't wrap and shows ellipsis if overflowed
+     */
+    singleLine?: boolean;
 }